refactor: change folder names: vex -> swap || varq -> forge
