+++ src/app/forge/page.tsx
@@ -3,8 +3,8 @@ import Image from "next/image";
 import CustomTab from "@/components/CustomTab";
 import { Card } from "@tremor/react";
 import React, { useEffect, useState } from "react";
-import VUSD_to_EMC from "@/components/varq/VUSD_to_EMC";
-import EMC_to_VUSD from "@/components/varq/EMC_to_VUSD";
+import VUSD_to_EMC from "@/components/forge/VUSD_to_EMC";
+import EMC_to_VUSD from "@/components/forge/EMC_to_VUSD";
 import VARQ_CONTRACT from "../../contracts/varq.json";
 import VTOKEN_CONTRACT from "../../contracts/vtoken.json";
 import { useAccount, useReadContract, useWriteContract } from "wagmi";
